requirements.txt
python-telegram-bot[webhooks]>=20.0
aiosqlite
httpx
aiocryptopay
apscheduler
python-dotenv
qrcode[pil]
